Move bindings to ./bindings
