+++ .gitignore
@@ -32,5 +32,5 @@ config.toml
 ./centrifuge-chain
 
 # Solidity
-/contracts
+/bindings
 /solidity
