+++ chains/ethereum/chain.go
@@ -7,7 +7,7 @@ import (
 	"fmt"
 
 	"github.com/ChainSafe/ChainBridgeV2/chains"
-	bridge "github.com/ChainSafe/ChainBridgeV2/contracts/Bridge"
+	bridge "github.com/ChainSafe/ChainBridgeV2/bindings/Bridge"
 	"github.com/ChainSafe/ChainBridgeV2/core"
 	"github.com/ChainSafe/ChainBridgeV2/crypto/secp256k1"
 	"github.com/ChainSafe/ChainBridgeV2/keystore"
