+++ chains/ethereum/connection.go
@@ -10,7 +10,7 @@ import (
 	"math/big"
 	"sync"
 
-	"github.com/ChainSafe/ChainBridgeV2/contracts/Bridge"
+	"github.com/ChainSafe/ChainBridgeV2/bindings/Bridge"
 	"github.com/ChainSafe/ChainBridgeV2/crypto/secp256k1"
 	"github.com/ChainSafe/log15"
